app/services/image_service.py
# Lógica principal: validação, importação, processamento 
import os
import shutil
import time
import threading
from pathlib import Path
from PIL import Image as PILImage
from sqlalchemy.orm import Session
from app.core.config import settings
# Removido: from app.crud.image_crud import ImageCRUD
import logging
from datetime import datetime

# Configure logging
logging.basicConfig(level=logging.INFO)
logger = logging.getLogger(__name__)

# Toda lógica de serviço relacionada à tabela Image removida conforme solicitado 